requirements.txt
alembic==1.15.2
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect==1.3.0
decorator==5.2.1
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
Flask==3.1.0
Flask-Migrate==4.1.0
flask-shell-ipython==0.5.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.3
nltk==3.9.1
numpy==2.2.5
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
playwright==1.51.0
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyperclip==1.9.0
python-dotenv==1.0.0
PyYAML==6.0.2
rank-bm25==0.2.2
rapidfuzz==2.15.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rpds-py==0.24.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
yt-dlp==2025.3.31
zipp==3.21.0
